+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Removed `Location` datatype in favour of `Building`
 - Refactor tests to run directly on tuple input to prevent unnecessary unpacking and repacking
 - Renamed usages of the word "room" to "location" in the codebase to better reflect the data represented
+- Added test cases for JSON parsing of Meeting data type in `backend-test/Database/TablesTests.hs`
 - Added test cases for JSON parsing of Time' data type in `backend-test/Database/TablesTests.hs`
 
 ## [0.7.2] - 2025-12-10